requirements.txt
playwright==1.41.1
pytest-playwright==0.4.4
python-dotenv==1.0.1
apscheduler==3.10.4
